the great restructure and refactor
