+++ bury.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
     <script type="module" src="/src/main.js" defer></script>
-    <script type="module" src="/src/pages/index.js" defer></script>
+    <script type="module" src="/src/pages/bury.js" defer></script>
     <link rel="stylesheet" href="/src/css/style.css" />
   </head>
