+++ graveyard.html
@@ -4,13 +4,10 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Graveyard - Productivity-Graveyard</title>
-    <link rel="stylesheet" href="./css/style.css" />
-    <script src="./src/components/header/header.js" defer></script>
-    <script
-      src="./src/components/projectCard/projectCard.js"
-      type="module"
-    ></script>
-    <script src="./src/components/footer/footer.js" defer></script>
+    <link rel="stylesheet" href="./src/css/style.css" />
+    <script type="module" src="/src/main.js" defer></script>
+    <script type="module" src="/src/pages/graveyard.js" defer></script>
+    <link rel="stylesheet" href="/src/css/style.css" />
   </head>
   <body>
     <header></header>
