--- js/main.js
@@ -1,3 +1,0 @@
-import { router } from "./router.js";
-
-router();
