+++ profile.html
@@ -4,10 +4,9 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Profile - Productivity-Graveyard</title>
-    <link rel="stylesheet" href="./css/style.css" />
-    <script src="./src/components/profile/profile.js" defer></script>
-    <script src="./src/components/projects/project.js" defer></script>
-    <script src="./src/components/header/header.js"></script>
+    <script type="module" src="/src/main.js" defer></script>
+    <script type="module" src="/src/pages/profile.js" defer></script>
+    <link rel="stylesheet" href="/src/css/style.css" />
   </head>
 
   <body>
@@ -41,5 +40,6 @@
         </div>
       </section>
     </main>
+    <footer></footer>
   </body>
 </html>
