+++ src/data/constants.js
@@ -1,0 +1,1 @@
+export const baseUrl = "localhost:3000";
