+++ src/main.js
@@ -1,0 +1,3 @@
+import { initUI } from "./components/ui/index.js";
+
+initUI();
