+++ termsofservice.html
@@ -8,20 +8,23 @@
       content="Terms of Use for Productivity Graveyard - Understanding your rights and responsibilities when using our service"
     />
     <title>Terms of Use | Productivity Graveyard</title>
-    <link rel="stylesheet" href="../../../../css/style.css" />
-    <link rel="stylesheet" href="../signup/terms.css" />
+    <script type="module" src="/src/main.js" defer></script>
+    <script type="module" src="/src/pages/index.js" defer></script>
+    <link rel="stylesheet" href="/src/css/style.css" />
   </head>
 
   <body>
     <div class="decorative-element moon"></div>
 
     <div class="container">
-      <header class="terms-header">
-        <h1>Terms of Use</h1>
-        <p>Read carefully before proceeding</p>
-      </header>
+      <header></header>
 
       <main class="terms-content">
+        <div class="terms-header">
+          <h1>Terms of Use</h1>
+          <p>Read carefully before proceeding</p>
+        </div>
+
         <p>
           Welcome to Productivity Graveyard. By accessing or using our website,
           you agree to be bound by these Terms of Use. Please read these terms
